Use gevent and remove the ioloop. Update tests.
It brings major changes in the design. Parallel execution is handled with
greenlets and a global poller. You MUST protect functions that may block.
Every node is registered to the poller, even clients. It is needed when the
client waits for the server's response.

When execution reaches the poller, it blocks in the poll() call. Then you need
to initialize everyting before switching the execution.