add a lock (semaphore) acquired by node send() and recv() to allow async handlers to share a socket
If a async handler call sendrecv(), it will wait before the _socket.recv() and
transfer the execution to the next greenlet. Then the socket is in "receiving"
state. The scheduler activates the poller. A message arrives. A new greenlet is
spawned with the async handler. When the execution reaches the _socket.send()
of the sendrecv(), it raises a exception "Operation cannot be accomplished in
current state".

The lock allows to ignore in which order greenlets are activated. As the first
_socket.recv() has acquired the lock, if the new greenlet is executed, it will
give back the execution when it tries to acquire the lock.