rewrite the way greenlets/nodes are spawned
Every creation of greenlet is now made through EventPoller.spawn.
It keeps track of spawned greenlet and when they exit (crashing or
exiting normally), actions can be made in the
:meth:`EventPoller.remove`.

When a greenlet stops with an unhandled exception, it cause every
greenlet in the process to stop too, leading to the end of the process
that was handling the greenlet. This can be useful when you have
something that watch processes lives.